requirements.txt
# requirements.txt
streamlit
openai-whisper
torch
transformers
sentence-transformers
faiss-cpu
textwrap
streamlit
git+https://github.com/openai/whisper.git
torch
ffmpeg